+++ src/FluxFlow.Components.Routing/Contracts/FlowWindowEmitReason.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Components.Routing.Contracts;
+
+public enum FlowWindowEmitReason
+{
+    Count = 0,
+    Time = 1,
+    Completion = 2
+}